requirements.txt
pandas>=2.0
pyyaml>=6.0
pyarrow>=15.0
spacy>=3.6
transformers>=4.40
torch>=2.0
pytest>=8.0
